+++ src/main/java/com/github/nicolasperuch/api/dto/VoteDto.java
@@ -1,15 +1,15 @@
 package com.github.nicolasperuch.api.dto;
 
 public class VoteDto {
-    private Long userId;
+    private Integer userId;
     private String userCpf;
     private boolean inFavor;
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
